requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
environs==11.0.0
frozenlist==1.4.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gspread==6.1.2
httplib2==0.22.0
idna==3.10
loguru==0.7.2
magic-filter==1.0.12
marshmallow==3.22.0
multidict==6.1.0
oauthlib==3.2.2
packaging==24.1
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
win32-setctime==1.1.0
yarl==1.12.1
